+++ Idle Defense/Assets/Scripts/PlayerBase/PlayerBaseUpgradeManager.cs
@@ -176,11 +176,11 @@ namespace Assets.Scripts.PlayerBase
             cost = RecursiveCost(baseCost, multiplier, currentLevel, outAmount);
         }
 
-        private ulong GetCost(PlayerBaseStatsInstance stats, PlayerUpgradeType type, int inAmount)
+        private float GetCost(PlayerBaseStatsInstance stats, PlayerUpgradeType type, int inAmount)
         {
             if (!_playerUpgrades.TryGetValue(type, out PlayerBaseUpgrade upgrade))
             {
-                return 0;
+                return 0f;
             }
 
             int currentLevel = upgrade.GetLevel(stats);
@@ -211,12 +211,12 @@ namespace Assets.Scripts.PlayerBase
             return amount;
         }
 
-        private ulong RecursiveCost(float baseCost, float multiplier, int level, int amount)
+        private float RecursiveCost(float baseCost, float multiplier, int level, int amount)
         {
             if (amount <= 0)
-                return 0;
+                return 0f;
 
-            ulong cost = (ulong)(baseCost * Mathf.Pow(multiplier, level));
+            float cost = baseCost * Mathf.Pow(multiplier, level);
             return cost + RecursiveCost(baseCost, multiplier, level + 1, amount - 1);
         }
