+++ Idle Defense/Assets/Scripts/UpgradeSystem/PlayerBaseUpgrade.cs
@@ -10,10 +10,10 @@ namespace Assets.Scripts.UpgradeSystem
         public Func<PlayerBaseStatsInstance, int> GetLevel;
         public Action<PlayerBaseStatsInstance, int> SetLevel;
         public Func<PlayerBaseStatsInstance, float> GetUpgradeAmount;
-        public Func<PlayerBaseStatsInstance, int> GetBaseCost;
+        public Func<PlayerBaseStatsInstance, float> GetBaseCost;
         public Func<PlayerBaseStatsInstance, float> GetMaxValue;
         public Func<PlayerBaseStatsInstance, float> GetMinValue;
-        public Func<PlayerBaseStatsInstance, int, ulong> GetCost;
+        public Func<PlayerBaseStatsInstance, int, float> GetCost;
         public Func<PlayerBaseStatsInstance, int> GetAmount;
         public Func<PlayerBaseStatsInstance, int, (string value, string bonus, string cost, string count)> GetDisplayStrings;
     }
